requirements.txt
# Core ADK dependencies
google-adk>=0.2.0
google-genai>=0.1.0
google-cloud-aiplatform>=1.38.0

# Agent dependencies  
rally-rest-api>=1.3.0
pygithub>=2.1.1
gitpython>=3.1.40
requests>=2.31.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
aiohttp>=3.9.1
redis>=5.0.1

# Code analysis tools
pylint>=3.0.3
radon>=6.0.1
lizard>=1.17.10

# Utilities
python-dotenv>=1.0.0
asyncio>=3.4.3
numpy>=1.26.2
pandas>=2.1.4

# Development tools
pytest>=7.4.3
pytest-asyncio>=0.21.1